+++ functions/index.js
@@ -523,6 +523,7 @@ exports.getReceipt = callable(webApi.getReceipt);
 exports.updateReceipt = callable(webApi.updateReceipt);
 exports.getReceiptImageUrls = callable(webApi.getReceiptImageUrls);
 exports.listProcessingFailures = callable(webApi.listProcessingFailures);
+exports.listNotifications = callable(webApi.listNotifications);
 exports.getProcessingFailureImageUrls = callable(webApi.getProcessingFailureImageUrls);
 exports.retryProcessing = callable(webApi.retryProcessing);
 exports.importTargetReceipts = callable(webApi.importTargetReceipts);